sf_simple/requirements.txt
Flask==2.3.3
PyPDF2==3.0.1
pdfplumber==0.9.0
Werkzeug==2.3.7
openai>=1.0.0
python-dotenv>=1.0.0
requests>=2.25.0